pyproject.toml
[project]
name = "graphql-sqlalchemy"
version = "0.6.1"
description = "Generate GraphQL Schemas from your SQLAlchemy models"
license = "MIT"
authors = [
    { name = "Guido Rainuzzo", email = "hi@guido.nyc" },
    { name = "Philipp A.", email = "flying-sheep@web.de" },
]
readme = "README.md"
urls.Homepage = "https://github.com/flying-sheep/graphql-sqlalchemy"
urls.Source = "https://github.com/flying-sheep/graphql-sqlalchemy"
keywords = ["graphql", "sqlalchemy"]
classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.9"
dependencies = [
    "graphql-core>=3",
    "SQLAlchemy>=2",
]

[project.optional-dependencies]
asyncio = [
    "SQLAlchemy[asyncio]",
]
test = [
    "pytest>=3.4",
    "pytest-cov>=2.9.0",
    "pytest-describe>=1.0.0",
    "exceptiongroup; python_version < '3.11'",
    "SQLAlchemy[asyncio]",
    "aiosqlite",
]
docs = [
    "pygments-graphql>=1.0.0",
    "sphinx>=3.5.3",
    "sphinx-rtd-theme>=0.5.1",
]

[project.entry-points."pygments.lexers"]
graphql = "pygments_graphql:GraphqlLexer"


[tool.hatch.envs.default]
python = "3.11"
dependencies = [
    "pygments", # pytest syntax highlighting
]

[[tool.hatch.envs.test.matrix]]
python = ["3.9", "3.10", "3.11"]
[tool.hatch.envs.test]
dependencies = ["pytest-watcher"]
features = ["test"]
[tool.hatch.envs.test.scripts]
run = "pytest -v {args}"
cov = "pytest -v --cov-report=term-missing --cov-report=xml --cov=graphql_sqlalchemy {args}"
watch = "ptw . -vv --cov-report=xml --cov=graphql_sqlalchemy {args}"

[tool.hatch.envs.docs]
features = ["docs"]
[tool.hatch.envs.docs.scripts]
build = "sphinx-build -M html docs/source docs/build"

[tool.pytest.ini_options]
xfail_strict = true
addopts = [
    '--import-mode=importlib',
    '--strict-markers',
    '-Werror',
]

[tool.mypy]
python_version = "3.11"
explicit_package_bases = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
select = [
    "E",
    "W",
    "F",
    "I",
    "UP",
    "PT",
]
[tool.ruff.isort]
required-imports = ["from __future__ import annotations"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
